refactor: counselor 도메인 구조 수정
