--- src/services/counselings/aggregates/counselors/infrastructures/adaptors/psql.counselors.repository.adaptor.ts
@@ -1,59 +1,0 @@
-import { UniqueEntityId } from "~shared/core/domain/UniqueEntityId";
-import { CounselorsEntity } from "~shared/core/infrastructure/entities/counsels/Counselors.entity";
-import { Counselors } from "~counselings/aggregates/counselors/domain/counselors";
-import { PsqlCounselorsMapper } from "~counselings/aggregates/counselors/infrastructures/adaptors/mapper/psql.counselors.mapper";
-import { CounselorsRepositoryPort, FindManyPropsInCounselorsRepository } from "~counselings/aggregates/counselors/infrastructures/counselors.repository.port";
-
-import { Injectable } from "@nestjs/common";
-import { InjectRepository } from "@nestjs/typeorm";
-import { FindOptionsWhere, Repository } from "typeorm";
-
-@Injectable()
-export class PsqlCounselorsRepositoryAdaptor implements CounselorsRepositoryPort {
-  constructor(
-    @InjectRepository(CounselorsEntity)
-    private readonly counselorsRepository: Repository<CounselorsEntity>,
-  ) {}
-
-  async create(counselor: Counselors): Promise<Counselors> {
-    const counselorsEntity = PsqlCounselorsMapper.toEntity(counselor);
-    await this.counselorsRepository.save(counselorsEntity);
-    return counselor;
-  }
-
-  async update(counselor: Counselors): Promise<Counselors> {
-    const counselorsEntity = PsqlCounselorsMapper.toEntity(counselor);
-    await this.counselorsRepository.update(counselorsEntity.id, counselorsEntity);
-    return counselor;
-  }
-
-  async findOne(counselorId: UniqueEntityId): Promise<Counselors | null> {
-    const counselorsEntity: CounselorsEntity | null = await this.counselorsRepository.findOne({
-      where: { id: counselorId.getString() },
-    });
-    if (!counselorsEntity) {
-      return null;
-    }
-    return PsqlCounselorsMapper.toDomain(counselorsEntity);
-  }
-
-  async findAll(): Promise<Counselors[]> {
-    const counselorsEntities = await this.counselorsRepository.find();
-    return counselorsEntities.map((counselorsEntity) => PsqlCounselorsMapper.toDomain(counselorsEntity)).filter((counselor) => counselor !== null);
-  }
-
-  async findMany(props: FindManyPropsInCounselorsRepository): Promise<Counselors[]> {
-    const findOptionsWhere: FindOptionsWhere<CounselorsEntity> = {};
-    if (props.name) {
-      findOptionsWhere.name = props.name;
-    }
-    if (props.toneId) {
-      findOptionsWhere.toneId = props.toneId.getString();
-    }
-
-    const counselorsEntities: CounselorsEntity[] = await this.counselorsRepository.find({
-      where: findOptionsWhere,
-    });
-    return counselorsEntities.map((counselorsEntity) => PsqlCounselorsMapper.toDomain(counselorsEntity)).filter((counselor) => counselor !== null);
-  }
-}
