+++ src/services/counselings/domains/counselors/counselors.criteria.ts
@@ -1,0 +1,6 @@
+import { UniqueEntityId } from "~shared/core/domain/UniqueEntityId";
+
+export type CounselorsCriteriaFindMany = {
+  name?: string;
+  toneId?: UniqueEntityId;
+};
